requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
mysql==0.0.3
mysql-connector-python==8.0.31
mysqlclient==2.1.1
Pillow==9.2.0
protobuf==3.20.1
psycopg2==2.9.4
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2022.4
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
